requirements.txt
ipykernel
pandas
biopython